move publications and methods to server
