+++ .meteor/versions
@@ -13,6 +13,7 @@ caching-compiler@1.1.11
 caching-html-compiler@1.1.2
 callback-hook@1.1.0
 check@1.3.1
+coffeescript@1.0.17
 ddp@1.4.0
 ddp-client@2.3.2
 ddp-common@1.4.0
@@ -52,6 +53,7 @@ mobile-status-bar@1.0.14
 modules@0.11.5
 modules-runtime@0.9.2
 momentjs:moment@2.21.0
+monbro:mongodb-mapreduce-aggregation@1.0.1
 mongo@1.4.5
 mongo-dev-server@1.1.0
 mongo-id@1.0.7
