+++ imports/ui/Auth.jsx
@@ -1,6 +1,6 @@
 import React, { Component, PropTypes } from 'react';
 import { withTracker } from 'meteor/react-meteor-data';
-import { Users } from '../api/classes.js';
+import { Users } from '../api/dbDefs.js';
 import {BrowserRouter, Route } from "react-router-dom";
 import { HTTP } from 'meteor/http';
