+++ imports/ui/Login.jsx
@@ -1,6 +1,6 @@
 import React, { Component, PropTypes } from 'react';
 import ReactDOM from 'react-dom';
-import { Reviews } from '../api/classes.js';
+import { Reviews } from '../api/dbDefs.js';
 import Admin from './Admin.jsx';
 import "./css/Login.css";
