+++ server/main.js
@@ -1,5 +1,4 @@
 import { Meteor } from 'meteor/meteor';
-import '../imports/api/classes.js';
 
 Meteor.startup(() => {
   // code to run on server at startup
